normalized how we were using tab layouts between different fragments.
